Add support for multiple arguments, bump version
I think that the `click` api made an error in the way they use
so many kwargs to control `click.option` behavior, which is part
of why I wrote `clickutil` in the first place. In the case of
multiple options, I think this is particularly bad; the multiple-ness
is part of the type.

So for clickutil, to add support for multiple options I added
the ability to use a dictionary, which can have a 'type' and
'multiple' entry. This allows you to declare a multiple input
(which in python-land will have type tuple) and the type of
the entries to that multiple input.

This change should be fully backward-compatible, no existing unit
tests required modification.